refactor: create `gui.rs` and `utils.rs`
